pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]

name="to-do-it"
version="0.0.1"
dependencies=[
    "pytest",
    "tkcalendar",
    "pillow"
]

[tool.setuptools.packages.find]

where=["fonte"]

[tool.pytest.ini_options]
pythonpath = ["."]  # Ou "src" se o seu código estiver dentro de uma pasta src
testpaths = ["tests"]